testing modifications

+++ cover/admin_settings.php
@@ -1,23 +1,34 @@
 <?php
   include_once 'header.php'
 ?>
-      <section>
-        <?php
-        if (isset($_SESSION["userEmail"])) {
-            ?>
-            <br>
-            <table>
-              <td><a href='create_user.php' class='button'>Create User</a></td>
-              <td><a href='delete_user.php' class='button'>Delete User</a></td>
-              <td><a href='import_teachers.php' class='button'>Import Teachers</a></td>
-              <td><a href='set_SLT.php' class='button'>Set SLT Teachers</a></td>
-            </table>
-            <?php
-        } else {
-            header("location: ./login.php");
-            exit();
-        }
-        ?>
+      <?php
+      if (isset($_SESSION["userEmail"])) {
+          if ($_SESSION["admin"] !== 0) {
+              ?>
+              <?php
+              if (isset($_SESSION["userEmail"])) {
+                  ?>
+                  <br>
+                  <table>
+                    <td><a href='create_user.php' class='button'>Create User</a></td>
+                    <td><a href='delete_user.php' class='button'>Delete User</a></td>
+                    <td><a href='import_teachers.php' class='button'>Import Teachers</a></td>
+                    <td><a href='set_SLT.php' class='button'>Set SLT Teachers</a></td>
+                  </table>
+              <?php
+              } else {
+                  header("location: ./login.php");
+                  exit();
+              }
+          } else {
+              header("location: ./index.php");
+              exit();
+          }
+      } else {
+          header("location: ./login.php");
+          exit();
+      }
+      ?>
       </section>
 <?php
   include_once 'footer.php'
